AM2 Open Source Commit
It's true!

AM2 is open source.  After an unfortunate loss of files that set me back
about a month of development, I have decided to put this project into
the hands of the community.  As such, a few bugs that were confirmed
complete are now in an either incomplete or partially complete state.

If you have questions as to the methodology of development or how to do
something, please create a new issue with the question tag (if you do
not have permission to do this, it will be assigned by a moderator).

If you are interested in becoming a moderator, please contact Mithion.