Change behaviour of AMFlower growth
Add canGrowOn() which is used in both the Grow component and the
spawning algorithm to spawn Tarma roots only in "dark places". Once the
flower is obtained, it can be placed on any light level, limited only by
canBlockStay().

Also increase the maximum height of AMFlowers to spawn from 126 to 254.